+++ rabbit_car/rabbit_car.ino
@@ -10,7 +10,7 @@
 
 // Define direction variables
 float MOTOR_SPEED = 1500;
-int SERVO_ANGLE = 90;
+float SERVO_ANGLE = 90;
 bool RUNNING = false;
 bool startRunTimer = true;
 bool manualControl = false;
